add ej 13

+++ 0x03-python-data_structures/13-is_palindrome.c
@@ -13,7 +13,7 @@ int _strlist(listint_t **head);
 int is_palindrome(listint_t **head)
 {
 	listint_t *salto = *head;
-	int sum_n = 0, i, len_list, mid, rest = 0;
+	long int sum_n = 0, i, len_list, mid, rest = 0;
 
 	if (head == NULL || *head == NULL)
 		return (1);
